moved typeinfo to its own file
