+++ server/types/type.go
@@ -25,7 +25,6 @@ import (
 
 	"github.com/cockroachdb/errors"
 	"github.com/dolthub/dolt/go/libraries/doltcore/schema/typeinfo"
-	dtypes "github.com/dolthub/dolt/go/store/types"
 	"github.com/dolthub/dolt/go/store/val"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/types"
@@ -613,7 +612,7 @@ func (t *DoltgresType) IoOutput(ctx *sql.Context, val any) (string, error) {
 // Or it can be pseudo category with name 'anyarray'.
 func (t *DoltgresType) IsArrayType() bool {
 	return (t.TypCategory == TypeCategory_ArrayTypes && t.Elem != id.NullType && t.Array == id.NullType) ||
-		(t.TypCategory == TypeCategory_PseudoTypes && t.ID.TypeName() == "anyarray")
+			(t.TypCategory == TypeCategory_PseudoTypes && t.ID.TypeName() == "anyarray")
 }
 
 // IsArrayCategory returns true if the type is of 'array' category.
@@ -1155,87 +1154,6 @@ func (t *DoltgresType) TypeInfo() typeinfo.TypeInfo {
 		Type: t,
 	}
 }
-
-type typeInfo struct {
-	Type     *DoltgresType
-	encoding val.Encoding
-}
-
-func (t typeInfo) Equals(other typeinfo.TypeInfo) bool {
-	ot, ok := other.(typeInfo)
-	return ok && t.Type.Equals(ot.Type) && t.encoding == ot.encoding
-}
-
-func (t typeInfo) NomsKind() dtypes.NomsKind {
-	return dtypes.ExtendedKind
-}
-
-func (t typeInfo) ToSqlType() sql.Type {
-	return t.Type
-}
-
-func (t typeInfo) Encoding() val.Encoding {
-	if t.encoding > 0 {
-		return t.encoding
-	}
-
-	// TODO: giant type switch for default encoding
-	switch t.Type.ID.TypeName() {
-	case "int2":
-		return val.Int16Enc
-	case "int4":
-		return val.Int32Enc
-	case "int8":
-		return val.Int64Enc
-	case "float4":
-		return val.Float32Enc
-	case "float8":
-		return val.Float64Enc
-	case "numeric", "decimal":
-		return val.DecimalEnc
-	case "bytea":
-		return val.BytesAdaptiveEnc
-	// TODO: use dolt JSON document encoding here
-	// case "json", "jsonb":
-	// 	return val.JSONAddrEnc
-	case "xid":
-		return val.Uint32Enc
-		// TODO: uuid is represented as a uuid.Uuid in doltgres, but dolt wants []byte for BytesAdaptiveEnc
-	// case "uuid":
-	// 	return val.BytesAdaptiveEnc
-	case "varchar":
-		if t.Type.attTypMod == -1 {
-			return val.StringAdaptiveEnc
-		}
-		return val.StringEnc
-	case "name", "char":
-		return val.StringEnc
-	case "bpchar", "text":
-		return val.StringAdaptiveEnc
-	default:
-		switch t.Type.MaxSerializedWidth() {
-		case sql.ExtendedTypeSerializedWidth_64K:
-			return val.ExtendedEnc
-		case sql.ExtendedTypeSerializedWidth_Unbounded:
-			return val.ExtendedAdaptiveEnc
-		default:
-			panic(fmt.Errorf("unknown extended type serialization width"))
-		}
-	}
-}
-
-func (t typeInfo) WithEncoding(enc val.Encoding) typeinfo.TypeInfo {
-	return typeInfo{
-		Type:     t.Type,
-		encoding: enc,
-	}
-}
-
-func (t typeInfo) String() string {
-	return fmt.Sprintf("TypeInfo(%s, encoding=%d)", t.Type.String(), t.encoding)
-}
-
-var _ typeinfo.TypeInfo = (*typeInfo)(nil)
 
 // TypeCastFunction is a function that takes a value of a particular kind of type, and returns it as another kind of type.
 // The targetType given should match the "To" type used to obtain the cast.
